Merge

+++ TP2/app/src/main/java/com/cedricnoiseux/tp2/User.java
@@ -34,7 +34,8 @@ public class User {
     }
 
     public static User getUser(String email, String groupName, boolean isAdmin,
-                        List<String> nomPreferences, int lastLocationX, int lastLocationY) {
+                        List<String> nomPreferences, float lastLocationX, float lastLocationY) {
+
         if (allUsers_ == null) {
             String csv = Utility.getAllLines("users.txt");
             csv = csv.trim();
@@ -55,6 +56,7 @@ public class User {
                 allUsers_.add(new User(e, g, a, act, pX, pY));
             }
         }
+
         Group g = Group.getGroup(groupName);
         List<ActivityType> a = new LinkedList<ActivityType>();
         for (String s : nomPreferences) {
@@ -68,11 +70,13 @@ public class User {
                 u = ret = new User(email, g, isAdmin, a, lastLocationX, lastLocationY );
             }
         }
+
         //rendu la, on sait que le groupe n'existe pas. On le cree
         if (!exists) {
             ret = new User(email, g, isAdmin, a, lastLocationX, lastLocationY );
             allUsers_.add(ret);
         }
+
         String o = "";
         for (User u : allUsers_) {
             o+= u.email_ + ";";
@@ -93,7 +97,4 @@ public class User {
         }
             return ret;
         }
-
-
-
 }
